+++ src/languages/markdown.js
@@ -105,12 +105,16 @@ export default function(hljs) {
       },
       // whatever else, lower relevance (might not be a link at all)
       {
-        begin: /\[.+?\]\(.*?\)/,
+        begin: /\[.*?\]\(.*?\)/,
         relevance: 0
       }
     ],
     returnBegin: true,
     contains: [
+      {
+        // empty strings for alt or link text
+        match: /\[(?=\])/
+      },
       {
         className: 'string',
         relevance: 0,